Changed the return type of NET_GetSocketAddressSize to an int

+++ Source/Headers/NetworkSocketAddress.h
@@ -12,7 +12,7 @@ typedef struct _tagSOCKET_ADDRESS
 /* Both the address and port should be in network byte order */
 int NET_CreateSocketAddress( PSOCKET_ADDRESS p_pSocketAddress,
 	Uint32 p_Address, Uint16 p_Port );
-size_t NET_GetSocketAddressSize( PSOCKET_ADDRESS p_pSocketAddress );
+int NET_GetSocketAddressSize( PSOCKET_ADDRESS p_pSocketAddress );
 
 #endif /* __TERMINAL_NETWORKSOCKETADDRESS_H__ */
